fix: fixed Estudiante so you can make POST, PUT and DELETE

+++ src/main/java/com/example/hogwarts/dto/AsignaturaDTO.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class AsignaturaDTO {
-    private Integer idAsignatura;
+    private Long idAsignatura;
     private String nombreAsignatura;
     private String aula;
     private Boolean asignaturaObligatoria;
