+++ src/main/java/com/example/hogwarts/dto/create/EstudianteCreateDTO.java
@@ -20,13 +20,13 @@ public class EstudianteCreateDTO {
     private String apellido;
 
     @NotNull(message = "El año del curso no puede estar vacío")
-    private Integer anyoCurso;
+    private long anyoCurso;
 
     @NotNull(message = "La fecha de nacimiento no puede estar vacía")
     @Past(message = "La fecha de nacimiento debe ser anterior a la fecha actual")
     private LocalDate fechaNacimiento;
 
-    private Integer idCasa;
+    private Long idCasa;
 
     @Valid
     @NotNull(message = "La mascota es obligatoria.")
