+++ src/main/java/com/example/hogwarts/repository/AsignaturaRepository.java
@@ -3,5 +3,5 @@ package com.example.hogwarts.repository;
 import com.example.hogwarts.model.Asignatura;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AsignaturaRepository extends JpaRepository<Asignatura, Integer> {
+public interface AsignaturaRepository extends JpaRepository<Asignatura, Long> {
 }
