+++ src/main/java/com/example/hogwarts/repository/CasaRepository.java
@@ -3,6 +3,6 @@ package com.example.hogwarts.repository;
 import com.example.hogwarts.model.Casa;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CasaRepository extends JpaRepository<Casa, Integer> {
+public interface CasaRepository extends JpaRepository<Casa, Long> {
     Casa getCasaByNombreCasa(String nombreCasa);
 }
