+++ src/main/java/com/example/hogwarts/repository/EstudianteRepository.java
@@ -3,5 +3,5 @@ package com.example.hogwarts.repository;
 import com.example.hogwarts.model.Estudiante;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EstudianteRepository extends JpaRepository<Estudiante, Integer> {
+public interface EstudianteRepository extends JpaRepository<Estudiante, Long> {
 }
