+++ src/main/java/com/example/hogwarts/repository/MascotaRepository.java
@@ -3,5 +3,5 @@ package com.example.hogwarts.repository;
 import com.example.hogwarts.model.Mascota;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MascotaRepository extends JpaRepository<Mascota, Integer> {
+public interface MascotaRepository extends JpaRepository<Mascota, Long> {
 }
